Convert builtin integer types to <stdint.h> types
Mechanical pass over all .c/.h files (compat/ included) mapping the
builtin integer types to fixed-width types under the original m68k/TIGCC
ABI where int == short == 16 bits and long == 32 bits:

  signed char                        -> int8_t
  unsigned char                      -> uint8_t
  short / int / signed short/int     -> int16_t
  unsigned / unsigned short/int      -> uint16_t
  long / long int / signed long      -> int32_t
  unsigned long / unsigned long int  -> uint32_t

Plain char is left untouched (it is the string/character type here).
int main(void) keeps its standard return type. Replacements skip
comments, string literals and char literals; per-file CRLF/LF endings
are preserved. #include <stdint.h> added wherever a fixed-width type is
now used and was not already available.

render.c: UINT_MAX -> UINT16_MAX so OldScore still initializes to 65535
(its m68k value) under the narrower uint16_t.

Builds clean; no new warnings.

+++ compat/gray.c
@@ -14,7 +14,7 @@
 // Note for the preprocessor's sake: a top-level comma in an EM_*_JS body would
 // split the macro arguments, since only parentheses are balanced. Keep commas
 // inside parentheses.
-EM_JS(void, pageflip, (const void *rgba, int w, int h), {
+EM_JS(void, pageflip, (const void *rgba, int16_t w, int16_t h), {
 	const canvas = Module.canvas;
 
 	// The canvas is sized from here rather than from the page, so the shell
@@ -55,9 +55,9 @@ EM_ASYNC_JS(void, pageflip_wait, (void), {
 	globalThis.__smPrevFrame = now - (elapsed % period);
 });
 
-void DelayNFrames(unsigned int frames)
+void DelayNFrames(uint16_t frames)
 {
-	for (int i = 0; i < frames; i++)
+	for (int16_t i = 0; i < frames; i++)
 		pageflip_wait();
 }
 
@@ -72,12 +72,12 @@ static bool dbuf_buf_no;
 static uint8_t renderbuf[LCD_HEIGHT][LCD_WIDTH][4];
 static void render(uint8_t *restrict light, uint8_t *restrict dark)
 {
-	for (int y = 0; y < LCD_HEIGHT; y++) {
-		for (int bx = 0; bx < LCD_LINE_BYTES; bx++) {
+	for (int16_t y = 0; y < LCD_HEIGHT; y++) {
+		for (int16_t bx = 0; bx < LCD_LINE_BYTES; bx++) {
 			uint8_t l = ~light[y * PLANE_STRIDE + bx];
 			uint8_t d = ~dark[y * PLANE_STRIDE + bx];
 
-			for (int j = 0; j < 8; j++) {
+			for (int16_t j = 0; j < 8; j++) {
 				uint8_t color =
 					(((l >> (7 - j)) & 1) * 0b01010101) |
 					(((d >> (7 - j)) & 1) * 0b10101010);
@@ -107,12 +107,12 @@ void GrayDBufInit(void *buf)
 	dbuf1 = buf;
 }
 
-void *GrayDBufGetActivePlane(short plane)
+void *GrayDBufGetActivePlane(int16_t plane)
 {
 	return dbuf_buf(dbuf_buf_no) + LCD_SIZE * !!plane;
 }
 
-void *GrayDBufGetHiddenPlane(short plane)
+void *GrayDBufGetHiddenPlane(int16_t plane)
 {
 	return dbuf_buf(!dbuf_buf_no) + LCD_SIZE * !!plane;
 }
